requirements.txt
aniso8601==9.0.1
backcall==0.2.0
biopython==1.79
CacheControl==0.12.11
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h5py==3.1.0
hdmedians==0.14.2
idna==3.3
importlib-metadata==4.8.1
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
kiwisolver==1.3.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.19.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.25
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
